Reviewed-on: https://gitea.apelsynca.pro/apelsynca/fragment-api/pulls/6

+++ server/src/fragment_transaction/schemas.py
@@ -21,5 +21,5 @@ class FragmentTransaction(TimestampedSchema):
 
 class ChartPoint(Schema):
     date: date
-    stars_spend: int
-    premium_spend: int
+    stars_spend: float
+    premium_spend: float
